+++ src/RK/computeVoronoiVolume.cc
@@ -279,7 +279,7 @@ std::vector<CellFaceFlag> extractFaceFlags(const GeomPolygon& cell,
       if (iclip < 0) {                                // Boundary clip (faceted boundary or void point)
         result.push_back(CellFaceFlag({(int)ifacet, -1, -1}));
       } else {                                        // Neighbor clip, iclip is the pair index in pairs
-        CHECK(iclip < pairs.size());
+        CHECK(iclip < (int)pairs.size());
         CHECK((pairs[iclip].i_list == nodeListi and pairs[iclip].i_node == i) or
               (pairs[iclip].j_list == nodeListi and pairs[iclip].j_node == i));
         if (pairs[iclip].i_list == nodeListi and pairs[iclip].i_node == i) {
@@ -323,7 +323,7 @@ std::vector<CellFaceFlag> extractFaceFlags(const GeomPolyhedron& cell,
       if (iclip < 0) {                                // Boundary clip (faceted boundary or void point)
         result.push_back(CellFaceFlag({(int)ifacet, -1, -1}));
       } else {                                        // Neighbor clip, iclip is the pair index in pairs
-        CHECK(iclip < pairs.size());
+        CHECK(iclip < (int)pairs.size());
         CHECK((pairs[iclip].i_list == nodeListi and pairs[iclip].i_node == i) or
               (pairs[iclip].j_list == nodeListi and pairs[iclip].j_node == i));
         if (pairs[iclip].i_list == nodeListi and pairs[iclip].i_node == i) {
